requirements.txt
Flask==3.1.3
Pillow
numpy
pandas
scikit-learn
scipy
matplotlib
seaborn
opencv-python-headless
kornia
python-json-logger
torch
torchvision
ultralytics
lap
deep-sort-realtime
python-bidi
paddlepaddle
paddleocr
pytesseract
pypdfium2

tensorflow
transformers
huggingface_hub
accelerate
